requirements.txt
flask
python-dotenv
mysql-connector-python
flask-wtf